README.txt
hello guys i'm Muheez, am here to work on frontend web development on hngi7 workplace...
So lets get started,

A front-end developer links together the world of design and technology, packaging up the utility of the back end in an inviting way for users to interact with. They take website design files and convert them into HTML, JavaScript (JS) and/or CSS code - the core elements of front-end development. This includes design/layout, content, buttons, images, navigation, and internal links.

By integrating markup languages, design, scripts and frameworks, front-end developers create the environment for everything that users see, click, and touch.

Everything a user sees, clicks, or uses to input or retrieve information on a website is the work of a front-end developer. This takes being both creative and tech-savvy. Their focus is on user experience, and the technology they implement hinges on how well it will accomplish that efficiency, speed, and smooth functionality. They also make sure that there are no errors or bugs on the front end, and that the design appears as it’s supposed to across various platforms and browsers.

Typically, a front-end developer’s responsibilities include:

    Prioritizing user experience
    Bringing a concept to life with HTML, CSS, and JavaScript
    Production and maintenance of websites and web application user interfaces
    Creating tools that enhance interaction with the site in any browser
    Implementing design for mobile sites
    Maintaining software workflow management
    Looking at SEO best practices
    Testing the site for usability and fixing any bugs